pyproject.toml
[tool.black]
line-length = 120
target-version = ["py310"]


[tool.poetry]
name = "opticapa"
version = "0.16.11"
description = "API services for Opticapa web app"
authors = ["Eric Santos <eric.santos@reseau.sncf.fr>"]

[tool.poetry.dependencies]
python = "^3.13"
geoalchemy2 = "0.14.3"
fastapi = "^0.115.0"
pydantic = "2.8.1"
networkx = "^3.0"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.4"
SQLAlchemy-Utils = "^0.41.0"
psycopg2-binary = "^2.9.3"
Shapely = "^2.0.1"
requests = "^2.30.0"
aiohttp = "^3.8.4"
asyncpg = "^0.30.0"
pytest = "^7.2.2"
pytest-env = "^1.1.0"
oauthlib = "^3.2.0"
pyjwt = "^2.7.0"
msgpack = "1.1.0"
ujson = "^5.4.0"
async-lru ="^2.0.4"
geojson-pydantic = "^1.0.1"
pytest-asyncio = "0.23.5"
pytest-mock = "^3.10.0"
pytest-dependency = "^0.5.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn="^0.25.0"
pre-commit = "^3.3.1"
opticapa_models = "0.11.5"
gunicorn = "^21.2.0"
coverage = "^7.2.3"
bandit = "1.7.5"
pytest-check = "^2.1.5"
openpyxl = "^3.1.2"
python-multipart = "^0.0.6"
cachetools = "5.3.1"
pyinstrument = "^4.6.0"
chardet = "^5.2.0"
pydantic-settings = "2.7.1"
python-dateutil = "^2.8.2"
python-json-logger = "^2.0.7"
pytz = "^2024.1"
optimisation = "0.2.3"
pydantic-core = "2.20.1"
annotated-types = "^0.7.0"
boto3= "1.35.18"
greenlet = "^3.2.2"
pbr = "^7.0.1"
opticapa-utils = "^0.1.3"

[tool.poetry.group.docs]
optional = true


[tool.poetry.group.tests]
optional = false

[tool.bandit]
targets = ["opticapa", "init_db"]
exclude_dirs = ["tests", "venv", "docker", "chart",  "docs", "certificates", "deployment"]
tests = ["B201", "B301"]

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.8"
mkdocstrings = "^0.19.0"
mkdocstrings-python = "^0.7.1"


[tool.poetry.group.tests.dependencies]
pytest = "^7.2.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = [
    "tests",
]

[tool.poetry.scripts]
init_db = "opticapa.shared.database.manage_db:init_database"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["__init__.py"]
